nasa_mixture_parsing no longer header only
