+++ src/Makefile.am
@@ -21,6 +21,7 @@ libantioch_la_SOURCES += parsing/src/blottner_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_ascii_parsing.C
 libantioch_la_SOURCES += parsing/src/chemkin_parser.C
+libantioch_la_SOURCES += parsing/src/nasa_mixture_parsing.C
 libantioch_la_SOURCES += parsing/src/sutherland_parsing.C
 libantioch_la_SOURCES += parsing/src/transport_species_parsing.C
 libantioch_la_SOURCES += parsing/src/xml_parser.C
